backend/pyproject.toml
[project]
name = "docmind-ai"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi",
  "uvicorn",
  "pydantic",
  "python-dotenv",
  "pydantic-settings",
  "langchain-openai",
  "pypdf",
  "python-multipart",
  "langchain>=1.2.17",
  "langchain-text-splitters",
  "faiss-cpu>=1.13.2",
  "numpy",
  "tiktoken>=0.12.0",
  "sqlalchemy>=2.0.49",
  "psycopg[binary]>=3.3.4",
  "celery[redis]>=5.4.0",
  "redis>=5.0.0",
  "bcrypt>=4.0.0",
  "python-jose[cryptography]>=3.3.0",
  "email-validator>=2.0.0",
  "alembic>=1.18.4",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["B008"]

[tool.ruff.lint.per-file-ignores]
"alembic/**" = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.11"
files = ["src"]
explicit_package_bases = true

[[tool.mypy.overrides]]
module = ["faiss", "faiss.*"]
ignore_missing_imports = true

[dependency-groups]
dev = [
    "celery-types>=0.26.0",
    "mypy>=2.1.0",
    "pytest>=8.4.2",
    "ruff>=0.15.12",
    "types-python-jose>=3.5.0.20260408",
]
